Improve test data fetch speed
The latter half of the method iterates on a specific list. The other list is ignored.
Mutate the sets if we want to remove moves. Then Remove all Giratina since we can't mutate form/item of the set.

Skips an entire re-parse for this one scenario, and the DistinctBy is better appearance/less code.